Explicitly specify the dtype of error tensor when it is initialized by zeros.

Previously if the dtype of input tensor is FP16, the error tensor is still created in FP32, although later it will be assigned by another FP16 tensor (`input_tensor_cp` - `input_tensor`).

This change will make the dtype of error tensor look more clear.

Additionally, also explicitly specify the dtype if rank-1 tensor buffer is empty.

Original PR issue: Investigate Applying PowerSGD to Communication Hook for Gradient Compression #47202

Differential Revision: [D26034988](https://our.internmc.facebook.com/intern/diff/D26034988/)

[ghstack-poisoned]

+++ torch/distributed/algorithms/ddp_comm_hooks/powerSGD_hook.py
@@ -232,7 +232,7 @@ def powerSGD_hook(state: PowerSGDState, bucket) -> torch.futures.Future:
     rank1_tensors_memory = (
         torch.cat([tensor.view(-1) for tensor in rank1_tensors])
         if rank1_tensors
-        else torch.tensor([], device=device)
+        else torch.tensor([], device=device, dtype=dtype)
     )
 
     # Step II: Handle high-rank tensors.
